Bug 18479 - Holds 'Placed on' column in opac-user.pl not sorting correctly
Two columns in the user's holds table in the OPAC do not sort correctly
because they are not marked up correctly to enable date sorting: 'Placed
on' and 'Expires on.' This patch corrects it.

This patch also removes a stray </td> which was causing validation
errors.

To test, apply the patch and log into the OPAC with an account which has
multiple holds with differnt hold dates and expiration dates. Confirm
that sorting on these columns works correctly.

Signed-off-by: Josef Moravec <josef.moravec@gmail.com>

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>

Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>