Bug 18799: XSLTresultsdisplay hides the icons
If item-level_itypes is turned off and XSLTResultsDisplay is turned on,
biblio-level itemtype images are not shown in staff client search
results. This patch corrects the problem.

This patch also adds some markup and CSS to the itemtype image to make
it look a litte cleaner and work better in result sets in which some
titles have images and some don't.

To test, apply the patch and view the staff client search results page
with various settings:

- XSLT turned on and off
- item-level_itypes turned on and off
- noItemTypeImages turned on and off

Confirm that item type images are shown or not shown correctly according
to each setting.

The difference between this patch and the other patch currently on this
bug is that this patch continues the practice of hiding biblio-level
itemtype information when item-level_itypes is ON.

Signed-off-by: Victor Grousset <victor.grousset@biblibre.com>

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>